+++ src/main/java/techcourse/myblog/presentation/controller/ArticleController.java
@@ -12,7 +12,7 @@ import techcourse.myblog.application.ArticleReadService;
 import techcourse.myblog.application.ArticleWriteService;
 import techcourse.myblog.application.CommentReadService;
 import techcourse.myblog.application.dto.ArticleDto;
-import techcourse.myblog.presentation.argument.LoginUser;
+import techcourse.myblog.presentation.support.LoginUser;
 
 import javax.validation.groups.Default;
 import java.util.List;
