+++ src/main/java/techcourse/myblog/presentation/support/LoginUser.java
@@ -1,9 +1,11 @@
-package techcourse.myblog.presentation.argument;
+package techcourse.myblog.presentation.support;
 
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import techcourse.myblog.domain.user.User;
 
 @Getter
+@EqualsAndHashCode
 public class LoginUser {
     private User user;
