+++ src/main/java/techcourse/myblog/presentation/support/exception/NotSignedInException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.presentation.argument.exception;
+package techcourse.myblog.presentation.support.exception;
 
 public class NotSignedInException extends RuntimeException {
     private static final String NOT_SIGNED_IN_EXCEPTION_MESSAGE = "로그인하지 않았습니다.";
